requirements.txt
streamlit
torch
torchvision
Pillow
ultralytics
gdown
numpy
pandas
matplotlib
tensorboard
scikit-learn